+++ packages/preprocess/index.cjs
@@ -46,16 +46,17 @@ const createDefaultProps = function(filename, componentDevelopmentMode, fileQuer
     let queryDeclarations = ''
     
     if(hasQueries(filename)) {
-        queryDeclarations = fileQueryIds?.filter(queryId => queryId.match('^([a-zA-Z_$][a-zA-Z0-9\d_$]*)$'))
-        .map(id => `let ${id} 
-        $: data, ${id} = data.${id};`)
-        .join('\n') || '';  
+        queryDeclarations = 
+        `
+        let {${fileQueryIds?.filter(queryId => queryId.match('^([a-zA-Z_$][a-zA-Z0-9\d_$]*)$')).map(id => id)} } = data;
+        $: ({${fileQueryIds?.filter(queryId => queryId.match('^([a-zA-Z_$][a-zA-Z0-9\d_$]*)$')).map(id => id)} } = data);
+        `
     } 
 
     let defaultProps = `
         import { page } from '$app/stores';
-        import { pageHasQueries } from '@evidence-dev/components/ui/stores';
-        import { setContext, getContext } from 'svelte';
+        import { pageHasQueries, routeHash } from '@evidence-dev/components/ui/stores';
+        import { setContext, getContext, beforeUpdate } from 'svelte';
         import BigLink from '${componentSource}/ui/BigLink.svelte';
         import Value from '${componentSource}/viz/Value.svelte';
         import BigValue from '${componentSource}/viz/BigValue.svelte';
@@ -70,7 +71,9 @@ const createDefaultProps = function(filename, componentDevelopmentMode, fileQuer
         import BarChart from '${componentSource}/viz/BarChart.svelte';
         import BubbleChart from '${componentSource}/viz/BubbleChart.svelte';
         import DataTable from '${componentSource}/viz/DataTable.svelte';
+        import Column from '${componentSource}/viz/Column.svelte';
         import LineChart from '${componentSource}/viz/LineChart.svelte';
+        import FunnelChart from "${componentSource}/viz/FunnelChart.svelte";
         import ScatterPlot from '${componentSource}/viz/ScatterPlot.svelte';
         import Histogram from '${componentSource}/viz/Histogram.svelte';
         import ECharts from '${componentSource}/viz/ECharts.svelte';
@@ -79,8 +82,8 @@ const createDefaultProps = function(filename, componentDevelopmentMode, fileQuer
         
         export let data = {};
         export let customFormattingSettings;
-        
-        let routeHash = '${routeHash}';
+
+        routeHash.set('${routeHash}');
 
         $: data, Object.keys(data).length > 0 ? pageHasQueries.set(true) : pageHasQueries.set(false);
 
@@ -89,6 +92,36 @@ const createDefaultProps = function(filename, componentDevelopmentMode, fileQuer
                 return customFormattingSettings.customFormats || [];
             }
         });
+        
+        const applyEvidenceTypes = function(data) {
+
+            let includedQueries = data.evidencemeta?.queries
+
+            if(includedQueries) {
+                // iterate through each query 
+                for(let i = 0; i < includedQueries.length; i++) {
+                    // for each of the query objects in data
+                    let query = data[includedQueries[i].id]
+                    let colTypes = data.evidencemeta?.queries[i].columnTypes
+                    // iterate through each row in the query
+                    for(let j = 0; j < query.length; j++) {
+                        // for each row in the query
+                        if(colTypes) {
+                            // include column types in the row object as a non enumerable property
+                            Object.defineProperty(query[j], '_evidenceColumnTypes', {
+                                enumerable: false,
+                                value: colTypes,
+                            });
+                        }
+                    }
+                }
+            }
+    
+        }
+    
+        beforeUpdate(() => {
+            applyEvidenceTypes(data)
+        })
 
         ${queryDeclarations}
         `