+++ sites/example-project/src/components/ui/Deployment/VercelDeploy.svelte
@@ -10,10 +10,10 @@
 <p>You'll need to set up a git repo before deploying to Vercel.</p>
 {:else }
 
-<h1>Deploying to Vercel</h1>
+<h2>Deploying to Vercel</h2>
 
 <ol>
-    <li><a href='https://vercel.com/new' target=_blank>Start a new Vercel project &rarr;</a></li>
+    <li><a href='https://vercel.com/new' target=_blank rel="noreferrer">Start a new Vercel project &rarr;</a></li>
     <li>Choose the repo containing this project</li>
     <li>Configure your project to match the settings below</li>
 
@@ -34,7 +34,7 @@
 
 <div class='setting-row'>
     <span class='setting'>Install Command</span>  
-    <div class='setting-value'><VariableCopy text={'build/'}/> </div>
+    <div class='setting-value'><VariableCopy text={'npm install'}/> </div>
 </div>