+++ sites/example-project/src/components/viz/Chart.svelte
@@ -74,7 +74,7 @@
     xGridlines = (xGridlines === "true" || xGridlines === true);
     export let xAxisLabels = true;
     xAxisLabels = (xAxisLabels === "true" || xAxisLabels === true);
-    export let sort = true; // sorts x values in case x is out of order in dataset (e.g., would create line chart that is out of order)
+    export let sort = false; // sorts x values in case x is out of order in dataset (e.g., would create line chart that is out of order)
     sort = (sort === "true" || sort === true);
 
     // Y axis:
@@ -160,6 +160,8 @@
 
     let missingCols = [];
 
+    let originalRun = true;
+
     // Error Handling:
 
     let inputCols = [];
@@ -203,7 +205,6 @@ $: {
 
         // If no y column(s) supplied, assume all number columns other than x are the y columns:
         if(!ySet) {
-            console.log('y fallback')
             uColNames = columnNames.filter(function(col){
                 return ![x, series, size].includes(col)
             });
@@ -249,6 +250,19 @@ $: {
             throw Error(new Intl.ListFormat().format(missingCols) + " are required");
         }
 
+        // Fix for stacked100 overwriting y variable. Bandaid fix - not a long-term solution:
+        if(stacked100 === true && y.includes("_pct") && originalRun === false){
+            if(typeof y === 'object'){
+                for(let i=0; i<y.length; i++){
+                    y[i] = y[i].replace("_pct", "")
+                }
+                originalRun = false;
+            } else {
+                y = y.replace("_pct", "")
+                originalRun = false;
+            }
+        }
+
         // Check the inputs supplied to the chart:
         if(x){inputCols.push(x)};
         if(y){
@@ -277,8 +291,10 @@ $: {
                 for(let i=0; i<y.length; i++){
                     y[i] = y[i] + '_pct'
                 }
+                originalRun = false;
             } else {
                 y = y + '_pct'
+                originalRun = false;
             }
 
             // Re-run column summary for new columns (not ideal):
@@ -328,6 +344,11 @@ $: {
                 getSortedData(data, y, false) 
                 : getSortedData(data, x, true) 
             : data;
+
+        // Always sort time axes by x - this prevents the lines from being drawn out of order
+        if(xDataType === "time"){
+            data = getSortedData(data, x, true);
+        }
     
     // ---------------------------------------------------------------------------------------
     // Standardize date columns
@@ -670,13 +691,15 @@ $: {
     }
 }
 
+$: data
+
 </script>
 
 
 {#if !error}
 
 <slot></slot>
-<ECharts config={$config} {height} {width}/>
+<ECharts config={$config} {height} {width} {data}/>
 
 {:else}