+++ packages/db-orchestrator/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @evidence-dev/db-orchestrator
 
+## 1.1.12
+
+### Patch Changes
+
+- 8894ea3: Prevent redundant queries being executed during build for parameterized pages
+
+## 1.1.11
+
+### Patch Changes
+
+- Updated dependencies [c013859]
+  - @evidence-dev/postgres@0.2.5
+  - @evidence-dev/redshift@0.0.4
+
+## 1.1.10
+
+### Patch Changes
+
+- Updated dependencies [a4371ed]
+  - @evidence-dev/duckdb@0.0.2
+
 ## 1.1.9
 
 ### Patch Changes