+++ packages/preprocess/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @evidence-dev/preprocess
 
+## 1.4.0
+
+### Minor Changes
+
+- 141dc5b: Improved DataTable component
+
+## 1.3.0
+
+### Minor Changes
+
+- e098028: Add FunnelChart component
+
+### Patch Changes
+
+- f5ec614: Ensure that precise evidence types are populated on all query objects
+
+## 1.2.6
+
+### Patch Changes
+
+- 79dca60: Fixes bug which prevented dev mode reloading on changes to queries in paramatarized pages
+- 755b1bb: Fix bug which caused bare query references to return undefined while page was loading
+
 ## 1.2.5
 
 ### Patch Changes