+++ sites/docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # evidence-docs
 
+## 1.2.0
+
+### Minor Changes
+
+- e098028: Add FunnelChart component
+
+## 1.1.0
+
+### Minor Changes
+
+- c68f658: Add maxWidth and minWidth style props to BigValue component. Default set for minWidth to give smaller BigValues more room.
+
 ## 1.0.4
 
 ### Patch Changes